Some cleanups releated to selection and component renaming
