+++ src/components/app/Form/Radio/RadioOption.module.scss
@@ -1,7 +1,3 @@
-.hintIconWrapper {
-  cursor: pointer;
-}
-
 .selectedBorderColor {
   border: 3px solid var(--primary-color);
 }
